feat(worker): extract worker back to its own process
